ml-service/requirements.txt
# ClimateCredit ML Service - Python Dependencies

fastapi==0.109.2
uvicorn[standard]==0.27.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.1
